[jwies/review-decision-chokepoints] review: split the gate hardening tests out (max-lines); drop stale imports
